+++ src/components/HTMLEngineProvider/BaseHTMLEngineProvider.js
@@ -10,7 +10,7 @@ import htmlRenderers from './HTMLRenderers';
 import * as HTMLEngineUtils from './htmlEngineUtils';
 import styles from '../../styles/styles';
 import fontFamily from '../../styles/fontFamily';
-import defaultViewProps from './defaultViewProps/defaultViewProps';
+import defaultViewProps from './defaultViewProps';
 
 const propTypes = {
     /** Whether text elements should be selectable */
